Create new Runtime Client crate
Extract the HTTP client logic into a new crate.
This allows other crates to share the same logic to
interact with the Runtime API.
