+++ lambda-runtime/src/lib.rs
@@ -6,21 +6,15 @@
 //! Create a type that conforms to the [`Handler`] trait. This type can then be passed
 //! to the the `lambda_runtime::run` function, which launches and runs the Lambda runtime.
 pub use crate::types::Context;
-use client::Client;
 use hyper::client::{connect::Connection, HttpConnector};
+use lambda_runtime_client::Client;
 use serde::{Deserialize, Serialize};
-use std::{
-    convert::{TryFrom, TryInto},
-    env, fmt,
-    future::Future,
-    panic,
-};
+use std::{convert::TryFrom, env, fmt, future::Future, panic};
 use tokio::io::{AsyncRead, AsyncWrite};
 use tokio_stream::{Stream, StreamExt};
 use tower_service::Service;
 use tracing::{error, trace};
 
-mod client;
 mod requests;
 #[cfg(test)]
 mod simulated;
@@ -30,14 +24,11 @@ mod types;
 use requests::{EventCompletionRequest, EventErrorRequest, IntoRequest, NextEventRequest};
 use types::Diagnostic;
 
-/// Error type that lambdas may result in
-pub type Error = Box<dyn std::error::Error + Send + Sync + 'static>;
+pub type Error = lambda_runtime_client::Error;
 
 /// Configuration derived from environment variables.
 #[derive(Debug, Default, Clone, PartialEq, Serialize, Deserialize)]
 pub struct Config {
-    /// The host and port of the [runtime API](https://docs.aws.amazon.com/lambda/latest/dg/runtimes-api.html).
-    pub endpoint: String,
     /// The name of the function.
     pub function_name: String,
     /// The amount of memory available to the function in MB.
@@ -54,7 +45,6 @@ impl Config {
     /// Attempts to read configuration from environment variables.
     pub fn from_env() -> Result<Self, Error> {
         let conf = Config {
-            endpoint: env::var("AWS_LAMBDA_RUNTIME_API").expect("Missing AWS_LAMBDA_RUNTIME_API env var"),
             function_name: env::var("AWS_LAMBDA_FUNCTION_NAME").expect("Missing AWS_LAMBDA_FUNCTION_NAME env var"),
             memory: env::var("AWS_LAMBDA_FUNCTION_MEMORY_SIZE")
                 .expect("Missing AWS_LAMBDA_FUNCTION_MEMORY_SIZE env var")
@@ -106,23 +96,8 @@ where
     }
 }
 
-#[non_exhaustive]
-#[derive(Debug, PartialEq)]
-enum BuilderError {
-    UnsetUri,
-}
-
 struct Runtime<C: Service<http::Uri> = HttpConnector> {
     client: Client<C>,
-}
-
-impl Runtime {
-    pub fn builder() -> RuntimeBuilder<HttpConnector> {
-        RuntimeBuilder {
-            connector: HttpConnector::new(),
-            uri: None,
-        }
-    }
 }
 
 impl<C> Runtime<C>
@@ -209,56 +184,6 @@ where
     }
 }
 
-struct RuntimeBuilder<C: Service<http::Uri> = hyper::client::HttpConnector> {
-    connector: C,
-    uri: Option<http::Uri>,
-}
-
-impl<C> RuntimeBuilder<C>
-where
-    C: Service<http::Uri> + Clone + Send + Sync + Unpin + 'static,
-    <C as Service<http::Uri>>::Future: Unpin + Send,
-    <C as Service<http::Uri>>::Error: Into<Box<dyn std::error::Error + Send + Sync>>,
-    <C as Service<http::Uri>>::Response: AsyncRead + AsyncWrite + Connection + Unpin + Send + 'static,
-{
-    pub fn with_connector<C2>(self, connector: C2) -> RuntimeBuilder<C2>
-    where
-        C2: Service<http::Uri> + Clone + Send + Sync + Unpin + 'static,
-        <C2 as Service<http::Uri>>::Future: Unpin + Send,
-        <C2 as Service<http::Uri>>::Error: Into<Box<dyn std::error::Error + Send + Sync>>,
-        <C2 as Service<http::Uri>>::Response: AsyncRead + AsyncWrite + Connection + Unpin + Send + 'static,
-    {
-        RuntimeBuilder {
-            connector,
-            uri: self.uri,
-        }
-    }
-
-    pub fn with_endpoint(self, uri: http::Uri) -> Self {
-        Self { uri: Some(uri), ..self }
-    }
-
-    pub fn build(self) -> Result<Runtime<C>, BuilderError> {
-        let uri = match self.uri {
-            Some(uri) => uri,
-            None => return Err(BuilderError::UnsetUri),
-        };
-        let client = Client::with(uri, self.connector);
-
-        Ok(Runtime { client })
-    }
-}
-
-#[test]
-fn test_builder() {
-    let runtime = Runtime::builder()
-        .with_connector(HttpConnector::new())
-        .with_endpoint(http::Uri::from_static("http://nomatter.com"))
-        .build();
-
-    runtime.unwrap();
-}
-
 fn incoming<C>(client: &Client<C>) -> impl Stream<Item = Result<http::Response<hyper::Body>, Error>> + Send + '_
 where
     C: Service<http::Uri> + Clone + Send + Sync + Unpin + 'static,
@@ -305,12 +230,8 @@ where
 {
     trace!("Loading config from env");
     let config = Config::from_env()?;
-    let uri = config.endpoint.clone().try_into().expect("Unable to convert to URL");
-    let runtime = Runtime::builder()
-        .with_connector(HttpConnector::new())
-        .with_endpoint(uri)
-        .build()
-        .expect("Unable to create a runtime");
+    let client = Client::builder().build().expect("Unable to create a runtime client");
+    let runtime = Runtime { client };
 
     let client = &runtime.client;
     let incoming = incoming(client);
@@ -319,4 +240,263 @@ where
 
 fn type_name_of_val<T>(_: T) -> &'static str {
     std::any::type_name::<T>()
+}
+
+#[cfg(test)]
+mod endpoint_tests {
+    use lambda_runtime_client::Client;
+    use crate::{
+        incoming,
+        requests::{
+            EventCompletionRequest, EventErrorRequest, IntoRequest, IntoResponse, NextEventRequest, NextEventResponse,
+        },
+        simulated,
+        types::Diagnostic,
+        Error, Runtime,
+    };
+    use http::{uri::PathAndQuery, HeaderValue, Method, Request, Response, StatusCode, Uri};
+    use hyper::{server::conn::Http, service::service_fn, Body};
+    use serde_json::json;
+    use simulated::DuplexStreamWrapper;
+    use std::{convert::TryFrom, env};
+    use tokio::{
+        io::{self, AsyncRead, AsyncWrite},
+        select,
+        sync::{self, oneshot},
+    };
+    use tokio_stream::StreamExt;
+
+    #[cfg(test)]
+    async fn next_event(req: &Request<Body>) -> Result<Response<Body>, Error> {
+        let path = "/2018-06-01/runtime/invocation/next";
+        assert_eq!(req.method(), Method::GET);
+        assert_eq!(req.uri().path_and_query().unwrap(), &PathAndQuery::from_static(path));
+        let body = json!({"message": "hello"});
+
+        let rsp = NextEventResponse {
+            request_id: "8476a536-e9f4-11e8-9739-2dfe598c3fcd",
+            deadline: 1_542_409_706_888,
+            arn: "arn:aws:lambda:us-east-2:123456789012:function:custom-runtime",
+            trace_id: "Root=1-5bef4de7-ad49b0e87f6ef6c87fc2e700;Parent=9a9197af755a6419",
+            body: serde_json::to_vec(&body)?,
+        };
+        rsp.into_rsp()
+    }
+
+    #[cfg(test)]
+    async fn complete_event(req: &Request<Body>, id: &str) -> Result<Response<Body>, Error> {
+        assert_eq!(Method::POST, req.method());
+        let rsp = Response::builder()
+            .status(StatusCode::ACCEPTED)
+            .body(Body::empty())
+            .expect("Unable to construct response");
+
+        let expected = format!("/2018-06-01/runtime/invocation/{}/response", id);
+        assert_eq!(expected, req.uri().path());
+
+        Ok(rsp)
+    }
+
+    #[cfg(test)]
+    async fn event_err(req: &Request<Body>, id: &str) -> Result<Response<Body>, Error> {
+        let expected = format!("/2018-06-01/runtime/invocation/{}/error", id);
+        assert_eq!(expected, req.uri().path());
+
+        assert_eq!(req.method(), Method::POST);
+        let header = "lambda-runtime-function-error-type";
+        let expected = "unhandled";
+        assert_eq!(req.headers()[header], HeaderValue::try_from(expected)?);
+
+        let rsp = Response::builder().status(StatusCode::ACCEPTED).body(Body::empty())?;
+        Ok(rsp)
+    }
+
+    #[cfg(test)]
+    async fn handle_incoming(req: Request<Body>) -> Result<Response<Body>, Error> {
+        let path: Vec<&str> = req
+            .uri()
+            .path_and_query()
+            .expect("PathAndQuery not found")
+            .as_str()
+            .split('/')
+            .collect::<Vec<&str>>();
+        match path[1..] {
+            ["2018-06-01", "runtime", "invocation", "next"] => next_event(&req).await,
+            ["2018-06-01", "runtime", "invocation", id, "response"] => complete_event(&req, id).await,
+            ["2018-06-01", "runtime", "invocation", id, "error"] => event_err(&req, id).await,
+            ["2018-06-01", "runtime", "init", "error"] => unimplemented!(),
+            _ => unimplemented!(),
+        }
+    }
+
+    #[cfg(test)]
+    async fn handle<I>(io: I, rx: oneshot::Receiver<()>) -> Result<(), hyper::Error>
+    where
+        I: AsyncRead + AsyncWrite + Unpin + 'static,
+    {
+        let conn = Http::new().serve_connection(io, service_fn(handle_incoming));
+        select! {
+            _ = rx => {
+                Ok(())
+            }
+            res = conn => {
+                match res {
+                    Ok(()) => Ok(()),
+                    Err(e) => {
+                        Err(e)
+                    }
+                }
+            }
+        }
+    }
+
+    #[tokio::test]
+    async fn test_next_event() -> Result<(), Error> {
+        let base = Uri::from_static("http://localhost:9001");
+        let (client, server) = io::duplex(64);
+
+        let (tx, rx) = sync::oneshot::channel();
+        let server = tokio::spawn(async {
+            handle(server, rx).await.expect("Unable to handle request");
+        });
+
+        let conn = simulated::Connector::with(base.clone(), DuplexStreamWrapper::new(client))?;
+        let client = Client::with(base, conn);
+
+        let req = NextEventRequest.into_req()?;
+        let rsp = client.call(req).await.expect("Unable to send request");
+
+        assert_eq!(rsp.status(), StatusCode::OK);
+        let header = "lambda-runtime-deadline-ms";
+        assert_eq!(rsp.headers()[header], &HeaderValue::try_from("1542409706888")?);
+
+        // shutdown server...
+        tx.send(()).expect("Receiver has been dropped");
+        match server.await {
+            Ok(_) => Ok(()),
+            Err(e) if e.is_panic() => Err::<(), Error>(e.into()),
+            Err(_) => unreachable!("This branch shouldn't be reachable"),
+        }
+    }
+
+    #[tokio::test]
+    async fn test_ok_response() -> Result<(), Error> {
+        let (client, server) = io::duplex(64);
+        let (tx, rx) = sync::oneshot::channel();
+        let base = Uri::from_static("http://localhost:9001");
+
+        let server = tokio::spawn(async {
+            handle(server, rx).await.expect("Unable to handle request");
+        });
+
+        let conn = simulated::Connector::with(base.clone(), DuplexStreamWrapper::new(client))?;
+        let client = Client::with(base, conn);
+
+        let req = EventCompletionRequest {
+            request_id: "156cb537-e2d4-11e8-9b34-d36013741fb9",
+            body: "done",
+        };
+        let req = req.into_req()?;
+
+        let rsp = client.call(req).await?;
+        assert_eq!(rsp.status(), StatusCode::ACCEPTED);
+
+        // shutdown server
+        tx.send(()).expect("Receiver has been dropped");
+        match server.await {
+            Ok(_) => Ok(()),
+            Err(e) if e.is_panic() => Err::<(), Error>(e.into()),
+            Err(_) => unreachable!("This branch shouldn't be reachable"),
+        }
+    }
+
+    #[tokio::test]
+    async fn test_error_response() -> Result<(), Error> {
+        let (client, server) = io::duplex(200);
+        let (tx, rx) = sync::oneshot::channel();
+        let base = Uri::from_static("http://localhost:9001");
+
+        let server = tokio::spawn(async {
+            handle(server, rx).await.expect("Unable to handle request");
+        });
+
+        let conn = simulated::Connector::with(base.clone(), DuplexStreamWrapper::new(client))?;
+        let client = Client::with(base, conn);
+
+        let req = EventErrorRequest {
+            request_id: "156cb537-e2d4-11e8-9b34-d36013741fb9",
+            diagnostic: Diagnostic {
+                error_type: "InvalidEventDataError".to_string(),
+                error_message: "Error parsing event data".to_string(),
+            },
+        };
+        let req = req.into_req()?;
+        let rsp = client.call(req).await?;
+        assert_eq!(rsp.status(), StatusCode::ACCEPTED);
+
+        // shutdown server
+        tx.send(()).expect("Receiver has been dropped");
+        match server.await {
+            Ok(_) => Ok(()),
+            Err(e) if e.is_panic() => Err::<(), Error>(e.into()),
+            Err(_) => unreachable!("This branch shouldn't be reachable"),
+        }
+    }
+
+    #[tokio::test]
+    async fn successful_end_to_end_run() -> Result<(), Error> {
+        let (client, server) = io::duplex(64);
+        let (tx, rx) = sync::oneshot::channel();
+        let base = Uri::from_static("http://localhost:9001");
+
+        let server = tokio::spawn(async {
+            handle(server, rx).await.expect("Unable to handle request");
+        });
+        let conn = simulated::Connector::with(base.clone(), DuplexStreamWrapper::new(client))?;
+
+        let client = Client::builder()
+            .with_endpoint(base)
+            .with_connector(conn)
+            .build()
+            .expect("Unable to build client");
+
+        async fn func(event: serde_json::Value, _: crate::Context) -> Result<serde_json::Value, Error> {
+            Ok(event)
+        }
+        let f = crate::handler_fn(func);
+
+        // set env vars needed to init Config if they are not already set in the environment
+        if env::var("AWS_LAMBDA_RUNTIME_API").is_err() {
+            env::set_var("AWS_LAMBDA_RUNTIME_API", "http://localhost:9001");
+        }
+        if env::var("AWS_LAMBDA_FUNCTION_NAME").is_err() {
+            env::set_var("AWS_LAMBDA_FUNCTION_NAME", "test_fn");
+        }
+        if env::var("AWS_LAMBDA_FUNCTION_MEMORY_SIZE").is_err() {
+            env::set_var("AWS_LAMBDA_FUNCTION_MEMORY_SIZE", "128");
+        }
+        if env::var("AWS_LAMBDA_FUNCTION_VERSION").is_err() {
+            env::set_var("AWS_LAMBDA_FUNCTION_VERSION", "1");
+        }
+        if env::var("AWS_LAMBDA_LOG_STREAM_NAME").is_err() {
+            env::set_var("AWS_LAMBDA_LOG_STREAM_NAME", "test_stream");
+        }
+        if env::var("AWS_LAMBDA_LOG_GROUP_NAME").is_err() {
+            env::set_var("AWS_LAMBDA_LOG_GROUP_NAME", "test_log");
+        }
+        let config = crate::Config::from_env().expect("Failed to read env vars");
+
+        let runtime = Runtime { client };
+        let client = &runtime.client;
+        let incoming = incoming(client).take(1);
+        runtime.run(incoming, f, &config).await?;
+
+        // shutdown server
+        tx.send(()).expect("Receiver has been dropped");
+        match server.await {
+            Ok(_) => Ok(()),
+            Err(e) if e.is_panic() => Err::<(), Error>(e.into()),
+            Err(_) => unreachable!("This branch shouldn't be reachable"),
+        }
+    }
 }
